Refactor and adaptations to integrate SDKin react application
