+++ apps/server/package.json
@@ -36,6 +36,7 @@
     "@nestjs/platform-express": "^10.3.9",
     "@nestjs/swagger": "^7.3.1",
     "aws-sdk": "^2.1640.0",
+    "axios": "^1.7.2",
     "class-transformer": "^0.5.1",
     "class-validator": "^0.14.1",
     "js-yaml": "^4.1.0",
