+++ sdks/typescript/package.json
@@ -1,7 +1,6 @@
 {
   "name": "@ethereal-pulse/typescript-sdk",
   "version": "0.0.1",
-  "type": "module",
   "license": "Apache-2.0",
   "author": "Diconium Digital Solutions",
   "description": "Ethereal Pulse Typescript SDK",
@@ -13,7 +12,7 @@
   },
   "scripts": {
     "test": "jest",
-    "build": "tsc",
+    "build": "tsc && tsc --module commonjs --outDir lib/cjs",
     "prepublishOnly": "pnpm run build",
     "format": "prettier --write \"src/**/*.ts\" \"test/**/*.ts\"",
     "lint": "eslint \"{src,apps,libs,test}/**/*.ts\" --fix"
@@ -25,9 +24,6 @@
     "email"
   ],
   "devDependencies": {
-    "jest": "^29.7.0",
-    "ts-jest": "^29.1.4",
-    "typescript": "^5.4.5",
     "@types/jest": "^29.5.12",
     "@types/node": "^20.14.2",
     "@typescript-eslint/eslint-plugin": "^6.7.4",
@@ -38,7 +34,16 @@
     "eslint-plugin-jsx-a11y": "^6.7.1",
     "eslint-plugin-prettier": "^5.1.3",
     "eslint-plugin-react": "^7.33.2",
-    "eslint-plugin-react-hooks": "^4.6.0"
+    "eslint-plugin-react-hooks": "^4.6.0",
+    "jest": "^29.7.0",
+    "ts-jest": "^29.1.4",
+    "typescript": "^5.4.5"
   },
-  "main": "dist/index.js"
+  "main": "./lib/cjs/index.js",
+  "module": "./lib/esm/index.js",
+  "types": "./lib/esm/index.d.ts",
+  "dependencies": {
+    "@types/react": "^18.2.20",
+    "axios": "^1.7.2"
+  }
 }
