+++ sdks/typescript/tsconfig.json
@@ -1,14 +1,33 @@
 {
   "compilerOptions": {
+    "outDir": "lib/esm",
     "target": "ES6",
-    "module": "commonjs",
-    "types": ["node", "jest"],
-    "strict": true,
+    "module": "ES6", // Consider changing to "ES6" if targeting modern bundlers
+    "jsx": "react",
+    "declaration": true,
+    "declarationMap": true,
+    "lib": [
+      "es6",
+      "dom",
+      "es2016",
+      "es2017"
+    ],
+    "moduleResolution": "node",
+    "noUnusedLocals": true,
+    "noUnusedParameters": true,
     "esModuleInterop": true,
-    "skipLibCheck": true,
-    "forceConsistentCasingInFileNames": true,
-    "outDir": "./dist"
+    "noImplicitReturns": true,
+    "noImplicitThis": true,
+    "noImplicitAny": true,
+    "strictNullChecks": true,
+    "allowSyntheticDefaultImports": true
   },
-  "include": ["./src/**/*.ts", "**/*.test.ts", "**/*.spec.ts"],
-  "exclude": ["node_modules", "dist"]
+  "include": [
+    "src/**/*"
+  ],
+  "exclude": [
+    "node_modules",
+    "**/*.test.ts",
+    "**/*.test.tsx"
+  ]
 }
